flow: create Flow-specific stdlib
This decouples Flow-specific stdlib functions to avoid import cycles.
