+++ pkg/flow/flow.go
@@ -55,6 +55,7 @@ import (
 	"github.com/go-kit/log/level"
 	"github.com/grafana/agent/pkg/flow/internal/controller"
 	"github.com/grafana/agent/pkg/flow/internal/dag"
+	"github.com/grafana/agent/pkg/flow/internal/stdlib"
 	"github.com/grafana/agent/pkg/flow/logging/v2"
 	"github.com/grafana/agent/pkg/flow/tracing"
 	"github.com/grafana/agent/pkg/river/vm"
@@ -247,7 +248,10 @@ func (c *Flow) LoadFile(file *File, args map[string]any) error {
 	}
 
 	argumentScope := &vm.Scope{
-		Parent: nil,
+		// The top scope is the Flow-specific stdlib.
+		Parent: &vm.Scope{
+			Variables: stdlib.Identifiers,
+		},
 		Variables: map[string]interface{}{
 			"argument": evaluatedArgs,
 		},
