+++ pkg/river/internal/stdlib/stdlib.go
@@ -5,9 +5,7 @@ import (
 	"encoding/json"
 	"os"
 
-	"github.com/grafana/agent/component/discovery"
 	"github.com/grafana/agent/pkg/river/internal/value"
-	"github.com/prometheus/prometheus/discovery/targetgroup"
 )
 
 // Identifiers holds a list of stdlib identifiers by name. All interface{}
@@ -73,34 +71,6 @@ var Identifiers = map[string]interface{}{
 		if err != nil {
 			return nil, err
 		}
-		return res, nil
-	},
-
-	"discovery_target_decode": func(in string) (interface{}, error) {
-		var targetGroups []*targetgroup.Group
-		if err := json.Unmarshal([]byte(in), &targetGroups); err != nil {
-			return nil, err
-		}
-
-		var res []discovery.Target
-
-		for _, group := range targetGroups {
-			for _, target := range group.Targets {
-
-				// Create the output target from group and target labels. Target labels
-				// should override group labels.
-				outputTarget := make(discovery.Target, len(group.Labels)+len(target))
-				for k, v := range group.Labels {
-					outputTarget[string(k)] = string(v)
-				}
-				for k, v := range target {
-					outputTarget[string(k)] = string(v)
-				}
-
-				res = append(res, outputTarget)
-			}
-		}
-
 		return res, nil
 	},
 }
