+++ pkg/river/vm/vm_stdlib_test.go
@@ -5,10 +5,8 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/grafana/agent/component/discovery"
 	"github.com/grafana/agent/pkg/river/parser"
 	"github.com/grafana/agent/pkg/river/vm"
-	"github.com/prometheus/common/model"
 	"github.com/stretchr/testify/require"
 )
 
@@ -37,69 +35,6 @@ func TestVM_Stdlib(t *testing.T) {
 
 			rv := reflect.New(reflect.TypeOf(tc.expect))
 			require.NoError(t, eval.Evaluate(nil, rv.Interface()))
-			require.Equal(t, tc.expect, rv.Elem().Interface())
-		})
-	}
-}
-
-func TestVM_Stdlib_Scoped(t *testing.T) {
-	tt := []struct {
-		name   string
-		input  string
-		scope  *vm.Scope
-		expect interface{}
-	}{
-		{
-			name:  "discovery_target_decode",
-			input: `discovery_target_decode(input)`,
-			scope: &vm.Scope{
-				Variables: map[string]interface{}{
-					"input": `[
-						{
-							"targets": ["host-a:12345", "host-a:12346"],
-							"labels": {
-								"foo": "bar"
-							}
-						},
-						{
-							"targets": ["host-b:12345", "host-b:12346"],
-							"labels": {
-								"hello": "world"
-							}
-						}
-					]`,
-				},
-			},
-			expect: []discovery.Target{
-				{
-					model.AddressLabel: "host-a:12345",
-					"foo":              "bar",
-				},
-				{
-					model.AddressLabel: "host-a:12346",
-					"foo":              "bar",
-				},
-				{
-					model.AddressLabel: "host-b:12345",
-					"hello":            "world",
-				},
-				{
-					model.AddressLabel: "host-b:12346",
-					"hello":            "world",
-				},
-			},
-		},
-	}
-
-	for _, tc := range tt {
-		t.Run(tc.name, func(t *testing.T) {
-			expr, err := parser.ParseExpression(tc.input)
-			require.NoError(t, err)
-
-			eval := vm.New(expr)
-
-			rv := reflect.New(reflect.TypeOf(tc.expect))
-			require.NoError(t, eval.Evaluate(tc.scope, rv.Interface()))
 			require.Equal(t, tc.expect, rv.Elem().Interface())
 		})
 	}
